+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Agreements accordion - Add a tooltip for the `New` and `Add` buttons. (UIEH-1424)
 * eHoldings package-title (resource) view: display Proxied URL. (UIEH-1419)
 * Fix title search is slow with many Packages filter options. (UIEH-1428)
+* Change permission `kb-ebsco.kb-credentials.item.get` to `kb-ebsco.kb-credentials.key.item.get` and `kb-ebsco.kb-credentials.uc.item.get` to `kb-ebsco.kb-credentials.uc.key.item.get`. (UIEH-1431)
 
 ## [9.1.1] (https://github.com/folio-org/ui-eholdings/tree/v9.1.1) (2024-03-24)